requirements.txt
arrow==1.2.2
articleDateExtractor==0.20
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
beautifulsoup4==4.9.3
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.2
cssselect==1.1.0
date-guesser==2.1.4
dateparser==1.1.0
docx==0.2.4
docx-mailmerge==0.5.0
docxcompose==1.3.2
elasticsearch==6.8.2
et-xmlfile==1.1.0
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.6.0
Flask==2.0.1
freezegun==1.2.0
future==0.18.2
gensim==4.0.1
googlesearch-python==1.1.0
h11==0.13.0
html5==0.0.9
htmldate==1.1.1
idna==2.8
importlib-metadata==4.11.2
itsdangerous==2.1.1
jieba3k==0.35.1
Jinja2==3.0.3
joblib==1.1.0
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.1.0
msedge-selenium-tools==3.141.4
newspaper3k==0.2.8
nltk==3.6.2
numpy==1.22.3
opencv-python==4.5.5.64
openpyxl==3.0.9
outcome==1.1.0
pandas==1.3.0
Pillow==9.0.1
pycodestyle==2.8.0
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.3.2
requests==2.25.1
requests-file==1.5.1
scipy==1.8.0
selenium==3.141.0
sgmllib3k==1.0.0
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
textblob==0.17.1
tinysegmenter==0.3
tldextract==3.1.0
toml==0.10.2
tqdm==4.63.0
trio==0.20.0
trio-websocket==0.9.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
vaderSentiment==3.3.2
webencodings==0.5.1
Werkzeug==2.0.3
wsproto==1.1.0
zipp==3.7.0
